Extract toExportSearchQuery helper
